// lightArduino/src/app/models/connections-schedule.ts
/**
 * @class ConnectionsSchedule
 */
export class ConnectionsSchedule {
    connectionName: string;
    isItDaily: boolean;
    configurationMaximumKilowattsPerDay: number;
    configurationMaximumKilowattsPerWeek: number;
    configurationMaximumKilowattsPerMonth: number;
    isItWeekly: boolean;
    isItMonthly: boolean;
    status: boolean;
}